requirements.txt
django
django-simple-captcha
requests
psycopg2
django-environ